+++ CHANGELOG.md
@@ -4,5 +4,11 @@
 
 ### Added
 
-- Added support for `inmanga` and `spyxfamily` non scheduled providers
-- Added support for `thetvdb` scheduled provider
+- feat: send telegram notification when a new non scheduled episode is found
+
+## [1.0.0] - 2023-03-05
+
+### Added
+
+- feat: add support for `inmanga` and `spyxfamily` non scheduled providers
+- feat: add support for `thetvdb` scheduled provider